# core/types: optimize bloom filters