This supersedes #21616 and closes #18092

The bloom filter was very suboptimal, and more or less broken. 


```
name                 old time/op    new time/op    delta
Bloom9-6               1.41µs ±17%    0.09µs ±10%  -93.75%  (p=0.008 n=5+5)
Bloom9Lookup-6         1.75µs ±15%    0.69µs ± 5%  -60.45%  (p=0.008 n=5+5)
CreateBloom/small-6    7.29µs ± 4%    2.86µs ± 0%  -60.76%  (p=0.016 n=5+4)
CreateBloom/large-6     701µs ±20%     280µs ± 2%  -59.96%  (p=0.008 n=5+5)

name                 old alloc/op   new alloc/op   delta
CreateBloom/small-6    7.50kB ± 0%    0.26kB ± 0%  -96.48%  (p=0.008 n=5+5)
CreateBloom/large-6     676kB ± 0%      26kB ± 0%  -96.21%  (p=0.008 n=5+5)

name                 old allocs/op  new allocs/op  delta
CreateBloom/small-6      58.0 ± 0%       9.0 ± 0%  -84.48%  (p=0.008 n=5+5)
CreateBloom/large-6     5.50k ± 0%     0.80k ± 0%  -85.44%  (p=0.008 n=5+5)
```